requirements.txt
agentlab==0.4.0
agentops==0.2.6
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.1
aws-requests-auth==0.4.3
backoff==2.2.1
beartype==0.12.0
beautifulsoup4==4.13.4
black==25.1.0
blacken-docs==1.19.1
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
browser-use==0.5.9
browsergym==0.14.1
browsergym-assistantbench==0.14.1
browsergym-core==0.14.1
browsergym-experiments==0.14.1
browsergym-miniwob==0.14.1
browsergym-visualwebarena==0.14.1
browsergym-webarena==0.14.1
browsergym-workarena==0.4.1
bubus==1.4.7
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
contexttimer==0.3.3
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
dask==2025.5.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
distlib==0.3.9
distributed==2025.5.1
distro==1.9.0
docstring_parser==0.16
english-words==2.0.1
evaluate==0.4.3
execnet==2.1.1
executing==2.2.0
Faker==37.3.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flaky==3.8.1
Flask==3.1.1
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.28.0
googleapis-common-protos==1.70.0
gradio==5.33.2
gradio_client==1.10.3
greenlet==3.2.3
groovy==0.1.2
groq==0.30.0
grpcio==1.73.0
gymnasium==1.1.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
identify==2.6.12
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
install_playwright==0.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy_loader==0.4
locket==1.0.0
lxml==5.4.0
markdown-it-py==3.0.0
markdown-pdf==1.5
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.12.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multion==1.3.8
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.41.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
nova-act==1.0.3679.0
numpy==2.2.6
oauthlib==3.3.1
ollama==0.5.1
openai==1.98.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.34.1
opentelemetry-exporter-prometheus==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
patchright==1.52.5
path==17.1.1
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.3.8
playwright==1.54.0
plotly==6.1.2
pluggy==1.6.0
portalocker==2.10.1
posthog==6.3.3
pre_commit==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.0
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyparsing==3.2.3
pypdf==5.9.0
pyperclip==1.9.0
PySocks==1.7.1
pytest==7.3.2
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.47.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.13
s3transfer==0.13.0
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
screeninfo==0.8.1
selenium==4.15.2
semantic-version==2.10.0
sentence-transformers==4.1.0
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
strands-agents==0.2.1
strands-agents-tools==0.1.8
sympy==1.14.0
tabulate==0.9.0
tblib==3.1.0
tenacity==8.5.0
termcolor==2.4.0
text-generation==0.7.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.21
tiktoken==0.9.0
tokenize_rt==6.2.0
tokenizers==0.21.1
tomlkit==0.13.3
toolz==1.0.0
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.4
types-requests==2.32.0.20250515
types-tqdm==4.67.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uuid7==0.1.0
uvicorn==0.34.2
virtualenv==20.31.2
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zict==3.0.0
zipp==3.23.0
zstandard==0.23.0
